requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.4.0
aiosqlite==0.22.1
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.10.4
arcade==3.3.3
async-timeout==4.0.3
asyncpg==0.29.0
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.4
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
magic-filter==1.0.12
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
narwhals==2.16.0
numpy==2.2.6
openai==2.28.0
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pillow==11.3.0
propcache==0.4.1
protobuf==6.33.5
pyaes==1.6.1
pyarrow==23.0.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.5.3
pydantic-extra-types==2.11.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
pydeck==0.9.1
pyglet==2.1.13
Pygments==2.19.2
pymunk==6.9.0
pyTelegramBotAPI==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.22
python-telegram-bot==22.5
pytiled_parser==2.2.9
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
referencing==0.37.0
requests==2.32.5
rich==14.3.2
rich-toolkit==0.18.1
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
sentry-sdk==2.52.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.52.1
streamlit==1.54.0
Telethon==1.42.0
tenacity==9.1.4
toml==0.10.2
tomli==2.4.0
tornado==6.5.4
tqdm==4.67.3
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
yarl==1.23.0
